logger was implemented, a resource folder was added and configured. Try catch block was implemented.It was also renamed a package to amazonWebTests.
