+++ src/main/java/pages/HomePage.java
@@ -1,5 +1,8 @@
 package pages;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.example.App;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -9,6 +12,7 @@ import org.openqa.selenium.support.ui.WebDriverWait;
 import java.time.Duration;
 
 public class HomePage {
+    private static final Logger LOGGER = LogManager.getLogger(App.class);
     private WebDriver driver;
     private By logo = By.cssSelector("[aria-label=\"Amazon\"]");
     private By signIn = By.cssSelector("#nav-link-accountList-nav-line-1");
@@ -34,10 +38,18 @@ public class HomePage {
         driver.findElement(burgerButton).click();
     }
 
-    public boolean isGreetingMessagePresent(){
-        WebDriverWait wait = new WebDriverWait(driver, Duration.ofMillis(5000));
-        return wait.until(ExpectedConditions.visibilityOf(
-                driver.findElement(greetingMessage))).isDisplayed();
+    public boolean isGreetingMessagePresent() {
+        LOGGER.info("Is_greeting_message_test is being executed");
+        boolean result = false;
+        try {
+            WebDriverWait wait = new WebDriverWait(driver, Duration.ofMillis(5000));
+           return result =  wait.until(ExpectedConditions.visibilityOf(
+                    driver.findElement(greetingMessage))).isDisplayed();
+        } catch (Exception e) {
+            LOGGER.error("Error displaying Greeting Message " + e);
+
+        }
+       return false;
     }
 
 }
